reduce the jobs handled by the dandelion pool
