--- p2p/dandelion-tower/src/pool/interface.rs
@@ -1,1 +1,0 @@
-/// The Service interface
